+++ capitalize_decorator.rb
@@ -1,0 +1,12 @@
+require './nameable'
+
+class CapitalizeDecorator < Nameable
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
