+++ student.rb
@@ -1,4 +1,5 @@
 require './person'
+
 class Student < Person
   def initialize(age, name = 'unknown', parent_permission: true)
     super(name, age, parent_permission)
