+++ teacher.rb
@@ -1,3 +1,5 @@
+require './person'
+
 class Teacher < Person
   def initialize(age, specialization, name = 'unknown', parent_permission: true)
     super(name, age, parent_permission)
