+++ trimmer_decorator.rb
@@ -1,0 +1,14 @@
+require './nameable'
+
+class TrimmerDecorator < Nameable
+  attr_accessor :nameable
+
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name[0...10]
+  end
+end
